+++ examples/ha_nodes.py
@@ -30,14 +30,30 @@ config = {
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
 oneview_client = OneViewClient(config)
-app_node_info = oneview_client.appliance_node_information
+ha_nodes = oneview_client.ha_nodes
 
-# Get node status information from appliance
-print("\nGet node status information from appliance:\n ")
-node_status = app_node_info.get_status()
-pprint(node_status.data)
+# Get all HA nodes from appliance
+print("\nGet all HA nodes from appliance:\n ")
+all_nodes = ha_nodes.get_all()
+for node in all_nodes:
+    print(" - {}".format(node['name']))
 
-# Get node version information from appliance
-print("\nGet node version information from appliance\n")
-node_version = app_node_info.get_version()
-pprint(node_version.data)
+# Get HA node by uri from appliance
+print("\nGet HA node by uri from appliance\n")
+node_by_uri = ha_nodes.get_by_uri(all_nodes[0]['uri'])
+pprint(node_by_uri.data)
+
+# update role of HA node
+data = {'role': 'Standby'}
+ha_node = node_by_uri.update(data)
+print("\n## Update HA node successfully!")
+pprint(ha_node.data)
+
+# Patch update role
+print("\nUpdate the HA node using patch")
+ha_node.patch(operation="replace", path="/role", value="Active")
+pprint(ha_node.data)
+
+# Delete HA node
+ha_node.delete()
+print("\n## Delete HA node successfully!")